+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 
 ## v0.0.33
 
+- **Enable star problem**
 - Fix tid data
 - Update record color